move itnerfaces to TestContext
