+++ packages/performance-tests/test/TransformerRegression.test.ts
@@ -15,7 +15,7 @@ import * as fs from "fs";
 import { BriefcaseDb, IModelHost, IModelHostConfiguration } from "@itwin/core-backend";
 import { Logger, LogLevel } from "@itwin/core-bentley";
 import { TransformerLoggerCategory } from "@itwin/imodel-transformer";
-import { getTestIModels } from "./TestContext";
+import { BriefcaseArgs, getTestIModels, ReporterInfo } from "./TestContext";
 import { filterIModels, initOutputFile, preFetchAsyncIterator } from "./TestUtils";
 import { NodeCliAuthorizationClient } from "@itwin/node-cli-authorization";
 import { BackendIModelsAccess } from "@itwin/imodels-access-backend";
@@ -28,28 +28,6 @@ import { getBranchName } from "./GitUtils";
 // cases
 import identityTransformer from "./cases/identity-transformer";
 import prepareFork from "./cases/prepare-fork";
-export interface ReporterInfo {
-  /* eslint-disable @typescript-eslint/naming-convention */
-  "Id": string;
-  "T-shirt size": string;
-  "Gb size": string;
-  "Branch Name": string;
-  "Federation Guid Saturation": number;
-  /* eslint-enable @typescript-eslint/naming-convention */
-}
-
-export interface ReporterEntry {
-  testSuite: string;
-  testName: string;
-  valueDescription: string;
-  value: number;
-  info?: ReporterInfo;
-}
-
-export interface BriefcaseArgs {
-  fileName: string;
-  briefcaseId: number;
-}
 
 const testCasesMap = new Map([
   ["identity transform", identityTransformer],
